refactor(profiles): move runtime functions from types to utils
Follow project convention by keeping shared/types/ for type definitions
only. Move conversion utilities and helper functions to shared/utils/:

- Create shared/utils/unified-account.ts for runtime functions
- Keep only types/interfaces in shared/types/unified-account.ts
- Update import in profile-scorer.ts to use new utils location

Functions moved:
- claudeProfileToUnified()
- apiProfileToUnified()
- isOAuthAccountId()
- isAPIAccountId()
- extractProfileId()
- toOAuthUnifiedId()
- toAPIUnifiedId()
- OAUTH_ID_PREFIX / API_ID_PREFIX constants

All 3055 frontend tests pass.
